issue #3131 refactor
